[change] Refactor ShareCard component to be reusable
