+++ src/components/personalPage/CollectedShares.jsx
@@ -1,8 +1,9 @@
 import React from 'react';
 import styled from 'styled-components';
+import ShareCard from '../common/ShareCard';
 
 const CollectedShares = () => {
-  return <div>CollectedShares</div>;
+  return <ShareCard />;
 };
 
 export default CollectedShares;
