--- src/components/personalPage/ShareCards.jsx
@@ -1,413 +1,0 @@
-import React, { useRef, useState } from 'react';
-import styled from 'styled-components';
-
-import { themeColor } from '../../utils/commonVariables';
-
-import { DialogOverlay, DialogContent } from '@reach/dialog';
-import '@reach/dialog/styles.css';
-
-import DateTimeSelector from '../common/DateTimeSelector';
-
-import Img from '../../images/restaurantPage/restaurant-8.jpg';
-import { ImSpoonKnife } from 'react-icons/im';
-import {
-  AiTwotoneStar,
-  AiTwotoneHeart,
-  AiFillCloseCircle,
-} from 'react-icons/ai';
-import { GrLocation } from 'react-icons/gr';
-import { BiCrown } from 'react-icons/bi';
-import { BsCalendarCheckFill } from 'react-icons/bs';
-
-const ShareCards = () => {
-  const uploadRef = useRef();
-  const [showEdit, setShowEdit] = useState(false);
-  const [showCalender, setShowCalendar] = useState(false);
-  const [file, setFile] = useState(null);
-
-  const openEditor = () => setShowEdit(true);
-  const closeEditor = () => setShowEdit(false);
-  const openCalendar = () => setShowCalendar(true);
-  const closeCalendar = () => setShowCalendar(false);
-
-  const previewImgUrl = file
-    ? URL.createObjectURL(file)
-    : 'https://react.semantic-ui.com/images/wireframe/image.png';
-
-  return (
-    <>
-      <SharesContext>
-        <ShareCard>
-          <ShareImg src={Img} />
-          <CardContent>
-            <ShareTitle>好吃的麵包</ShareTitle>
-            <DeleteButton />
-            <CardRow>
-              <CardItem>
-                <ShareNameIcon />
-                <ShareUseName>麵包超人</ShareUseName>
-              </CardItem>
-              <CardItem>
-                <Star />
-                <Rating>5</Rating>
-              </CardItem>
-              <Heart />
-            </CardRow>
-            <CardRow>
-              <CardItem>
-                <PlaceIcon />
-                <Location>台北 板橋</Location>
-              </CardItem>
-              <GetButton onClick={openEditor}>編輯</GetButton>
-            </CardRow>
-          </CardContent>
-        </ShareCard>
-      </SharesContext>
-      <DialogOverlay isOpen={showEdit} onDismiss={closeEditor}>
-        <DialogContent
-          style={{
-            position: 'relative',
-            border: 'solid 1px lightBlue',
-            borderRadius: '10px',
-          }}
-        >
-          <PopClose onClick={closeEditor} />
-          <PopTitleContainer>
-            <CrownIcon />
-            <PopTitle>好吃的麵包</PopTitle>
-          </PopTitleContainer>
-          <PopContent>
-            <PopRow>
-              <FoodLabel>食物名稱</FoodLabel>
-              <FoodName />
-            </PopRow>
-            <PopRow>
-              <QuantityLabel>數量</QuantityLabel>
-              <Quantity />
-            </PopRow>
-            <PopRow>
-              <DateTimeLabel>日期及時間</DateTimeLabel>
-              <DateTime>2021-10-15 20:00</DateTime>
-              <Calendar onClick={openCalendar} />
-            </PopRow>
-            <PopRow>
-              <PopPlaceLabel>地點</PopPlaceLabel>
-              <PopPlace>台北．內湖</PopPlace>
-              <PopPlaceIcon />
-            </PopRow>
-            <PopRow>
-              <FoodImgLabel>食物照片</FoodImgLabel>
-              <ImgUpload ref={uploadRef} htmlFor="image-upload">
-                上傳
-              </ImgUpload>
-              <UploadBtn
-                type="file"
-                id="image-upload"
-                onChange={(e) => setFile(e.target.files[0])}
-              />
-            </PopRow>
-            <PreviewImg src={previewImgUrl} />
-            <SubmitBtn>確認更新</SubmitBtn>
-          </PopContent>
-        </DialogContent>
-      </DialogOverlay>
-      <DialogOverlay isOpen={showCalender} onDismiss={closeCalendar}>
-        <DialogContent
-          style={{
-            position: 'relative',
-            border: 'solid 1px lightBlue',
-            borderRadius: '10px',
-          }}
-        >
-          <PopClose onClick={closeCalendar} />
-          <DateTimeSelector />
-        </DialogContent>
-      </DialogOverlay>
-    </>
-  );
-};
-
-const SharesContext = styled.div`
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
-  grid-template-rows: auto;
-  gap: 2rem;
-  padding: 3vw 4rem;
-  @media screen and (max-width: 700px) {
-    grid-template-columns: 1fr;
-    gap: 1rem;
-  }
-`;
-
-const ShareCard = styled.div`
-  display: flex;
-  border: 1px solid black;
-  border-radius: 10px;
-`;
-
-const ShareImg = styled.img`
-  max-width: 15vw;
-  border-radius: 10px 0 0 10px;
-  @media screen and (max-width: 700px) {
-    max-width: 30vw;
-  }
-  @media screen and (max-width: 460px) {
-    max-width: 25vw;
-  }
-`;
-
-const CardRow = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  flex-wrap: nowrap;
-  font-size: 1.5vw;
-  &:nth-child(2) {
-    margin-bottom: 1vw;
-  }
-
-  @media screen and (max-width: 800px) {
-    &:nth-child(3) {
-      flex-wrap: wrap;
-    }
-  }
-
-  @media screen and (max-width: 700px) {
-    font-size: 2.6vw;
-    &:nth-child(2) {
-      margin-bottom: 1vw;
-    }
-  }
-`;
-
-const CardItem = styled.div`
-  display: flex;
-  flex-wrap: nowrap;
-`;
-
-const CardContent = styled.div`
-  display: flex;
-  position: relative;
-  flex-direction: column;
-  justify-content: space-between;
-  padding: 2vw 1.5vw;
-  flex-grow: 1;
-  @media screen and (max-width: 700px) {
-    padding: 3.5vw 3vw;
-  }
-`;
-
-const ShareTitle = styled.h2`
-  font-size: 2vw;
-  margin-bottom: 2vw;
-  @media screen and (max-width: 700px) {
-    font-size: 3vw;
-    margin-bottom: 2vw;
-  }
-  @media screen and (max-width: 460px) {
-    font-size: 3.5vw;
-    margin-bottom: 3vw;
-  }
-`;
-
-const ShareNameIcon = styled(ImSpoonKnife)`
-  fill: ${themeColor.iconColor};
-  margin-right: 1vw;
-  width: 1.8vw;
-  height: 1.8vw;
-  @media screen and (max-width: 700px) {
-    width: 2.6vw;
-    height: 2.6vw;
-  }
-`;
-
-const ShareUseName = styled.span``;
-
-const StyledColse = styled(AiFillCloseCircle)`
-  fill: lightblue;
-  background-color: blue;
-  border-radius: 50%;
-  opacity: 0.8;
-  cursor: pointer;
-`;
-
-const DeleteButton = styled(StyledColse)`
-  position: absolute;
-  top: -1.2vw;
-  right: -1.2vw;
-  width: 3vw;
-  height: 3vw;
-
-  @media screen and (max-width: 700px) {
-    width: 4vw;
-    height: 4vw;
-    top: -2vw;
-    right: -2vw;
-  }
-`;
-
-const Star = styled(AiTwotoneStar)`
-  fill: orchid;
-  margin-right: 1vw;
-  width: 1.8vw;
-  height: 1.8vw;
-  @media screen and (max-width: 700px) {
-    width: 2.6vw;
-    height: 2.6vw;
-  }
-`;
-
-const Rating = styled.span``;
-
-const Heart = styled(AiTwotoneHeart)`
-  fill: red;
-  width: 1.8vw;
-  height: 1.8vw;
-  @media screen and (max-width: 700px) {
-    width: 2.6vw;
-    height: 2.6vw;
-  }
-`;
-
-const PlaceIcon = styled(GrLocation)`
-  margin-right: 1vw;
-  width: 1.8vw;
-  height: 1.8vw;
-`;
-
-const Location = styled.span``;
-
-const GetButton = styled.div`
-  border: 1px solid ${themeColor.outLineColor};
-  padding: 0.5rem 1rem;
-  border-radius: 8px;
-  background-color: #2a9d8f;
-  color: white;
-  cursor: pointer;
-
-  @media screen and (max-width: 860px) {
-    padding: 0.3rem 0.5rem;
-  }
-
-  @media screen and (max-width: 800px) {
-    margin-top: 0.5rem;
-  }
-`;
-
-const PopClose = styled(StyledColse)`
-  position: absolute;
-  top: 2vw;
-  right: 2vw;
-  width: 3vw;
-  height: 3vw;
-  cursor: pointer;
-`;
-
-const PopTitleContainer = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding-bottom: 1vw;
-  border-bottom: 1px solid lightskyblue;
-`;
-
-const CrownIcon = styled(BiCrown)`
-  fill: lightskyblue;
-  width: 3vw;
-  height: 3vw;
-  margin-right: 2vw;
-`;
-
-const PopTitle = styled.div`
-  font-size: 2.5vw;
-`;
-
-const PopContent = styled.div`
-  display: flex;
-  flex-direction: column;
-  padding: 2vw 1.5vw;
-`;
-
-const PopRow = styled.div`
-  display: flex;
-  width: 100%;
-  align-items: center;
-  font-size: 1.5vw;
-  margin-bottom: 2vw;
-`;
-
-const FoodLabel = styled.label`
-  width: 9vw;
-`;
-
-const FoodName = styled.input`
-  flex-grow: 1;
-`;
-
-const QuantityLabel = styled.label`
-  width: 9vw;
-`;
-
-const Quantity = styled.input``;
-
-const DateTimeLabel = styled.label`
-  width: 9vw;
-`;
-
-const DateTime = styled.span`
-  margin-right: 1vw;
-`;
-
-const Calendar = styled(BsCalendarCheckFill)`
-  width: 2vw;
-  height: 2vw;
-  fill: lightseagreen;
-  cursor: pointer;
-`;
-
-const PopPlaceLabel = styled.label`
-  width: 9vw;
-`;
-
-const PopPlace = styled.span`
-  margin-right: 1vw;
-`;
-
-const PopPlaceIcon = styled(GrLocation)`
-  width: 2vw;
-  height: 2vw;
-`;
-
-const FoodImgLabel = styled.label`
-  width: 9vw;
-`;
-
-const ImgUpload = styled.label`
-  border: 1px solid lightslategrey;
-  border-radius: 5px;
-  background-color: lightskyblue;
-  padding: 0.5vw;
-  cursor: pointer;
-`;
-
-const UploadBtn = styled.input`
-  display: none;
-`;
-
-const PreviewImg = styled.img`
-  border-radius: 10px;
-  margin-bottom: 2vw;
-`;
-
-const SubmitBtn = styled.button`
-  flex-grow: 1;
-  border: none;
-  border-radius: 5px;
-  background-color: lightskyblue;
-  color: white;
-  cursor: pointer;
-  padding: 1vw;
-  letter-spacing: 0.5vw;
-`;
-
-export default ShareCards;
