+++ src/components/personalPage/toReceive.jsx
@@ -1,8 +1,9 @@
 import React from 'react';
 import styled from 'styled-components';
+import ShareCard from '../common/ShareCard';
 
 const ToReceive = () => {
-  return <div>ToReceive</div>;
+  return <ShareCard />;
 };
 
 export default ToReceive;
